pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.58.2",
    "docx>=0.2.4",
    "google-generativeai>=0.8.5",
    "openai>=1.97.0",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "reportlab>=4.4.2",
    "requests>=2.32.4",
    "streamlit>=1.47.0",
]
